+++ openapi-codegen/core/src/main/scala/sttp/tapir/codegen/endpoints/EndpointGenerator.scala
@@ -3,25 +3,17 @@ package sttp.tapir.codegen.endpoints
 import io.circe.Json
 import sttp.tapir.codegen.dedup.PackageReuseContext
 import sttp.tapir.codegen.endpoints.SimpleTypes.mapSchemaSimpleTypeToType
+import sttp.tapir.codegen.endpoints.InAndOutComponents._
 import sttp.tapir.codegen.json.JsonSerdeLib.JsonSerdeLib
 import sttp.tapir.codegen.openapi.models._
 import sttp.tapir.codegen.openapi.models.GenerationDirectives._
-import sttp.tapir.codegen.openapi.models.OpenapiModels.{
-  OpenapiDocument,
-  OpenapiParameter,
-  OpenapiPath,
-  OpenapiRequestBodyContent,
-  OpenapiRequestBodyDefn,
-  OpenapiResponseContent,
-  OpenapiResponseDef
-}
+import sttp.tapir.codegen.openapi.models.OpenapiModels.{OpenapiDocument, OpenapiParameter, OpenapiPath}
 import sttp.tapir.codegen.openapi.models.OpenapiSchemaType._
 import sttp.tapir.codegen.security.{SecurityDefn, SecurityGenerator, SecurityWrapperDefn}
 import sttp.tapir.codegen.util.ErrUtils.bail
 import sttp.tapir.codegen.util.NameHelpers.{indent, strippedToCamelCase}
-import sttp.tapir.codegen.util.{JavaEscape, Location, NameHelpers}
-import sttp.tapir.codegen.validation.{ValidationDefns, ValidationGenerator}
-import sttp.tapir.codegen.xml.XmlSerdeGenerator
+import sttp.tapir.codegen.util.{JavaEscape, Location}
+import sttp.tapir.codegen.validation.ValidationDefns
 import sttp.tapir.codegen.xml.XmlSerdeLib
 import sttp.tapir.codegen.xml.XmlSerdeLib.XmlSerdeLib
 
@@ -38,12 +30,6 @@ case class EndpointTypes(security: Seq[String], in: Seq[String], err: Seq[String
 }
 case class GeneratedEndpoint(name: String, definition: String, maybeInlineDefns: Option[String], types: EndpointTypes)
 case class GeneratedEndpointsForFile(maybeFileName: Option[String], generatedEndpoints: Seq[GeneratedEndpoint])
-
-object Position extends Enumeration {
-  val Request, Response, Err = Value
-  type Position = Value
-}
-import sttp.tapir.codegen.endpoints.Position._
 
 case class EndpointDetails(
     jsonParamRefs: Set[String],
@@ -86,35 +72,8 @@ case class EndpointDefs(
     details: EndpointDetails
 )
 
-case class UrlMapResponse(
-    pathDecl: String,
-    pathTypes: Seq[String],
-    maybeSecurityPath: Option[(String, Seq[String])],
-    inlineUrlDefs: Seq[String]
-)
-
 class EndpointGenerator {
-  private def combine(inlineDefn1: Option[String], inlineDefn2: Option[String], separator: String = "\n") =
-    (inlineDefn1, inlineDefn2) match {
-      case (None, None)               => None
-      case (Some(defn), None)         => Some(defn)
-      case (None, Some(defn))         => Some(defn)
-      case (Some(defn1), Some(defn2)) => Some(defn1 + separator + defn2)
-    }
   private[codegen] def allEndpoints: String = "generatedEndpoints"
-
-  private def capabilityImpl(streamingImplementation: StreamingImplementation): String = streamingImplementation match {
-    case Akka            => "sttp.capabilities.akka.AkkaStreams"
-    case FS2(effectType) => s"sttp.capabilities.fs2.Fs2Streams[$effectType]"
-    case Pekko           => "sttp.capabilities.pekko.PekkoStreams"
-    case Zio             => "sttp.capabilities.zio.ZioStreams"
-  }
-  private def capabilityType(streamingImplementation: StreamingImplementation): String = streamingImplementation match {
-    case FS2(effectType) => s"fs2.Stream[$effectType, Byte]"
-    case x               => s"${capabilityImpl(x)}.BinaryStream"
-  }
-  // These types all use 'eager' schemas, except for '*/*', which we default to eager for convenience but which has no schema mappings
-  private val eagerTypes = Set("application/json", "application/xml", "text/plain", "text/html", "multipart/form-data", "*/*")
 
   def endpointDefs(
       doc: OpenapiDocument,
@@ -216,7 +175,7 @@ class EndpointGenerator {
           val name = m.name(p.url)
           val isReused = packageReuse.reusedEndpointNames.contains(name)
           def maybeName = m.operationId.map(n => s"""\n  .name("${JavaEscape.escapeString(n)}")""").getOrElse("")
-          val UrlMapResponse(pathDecl, pathTypes, maybeSecurityPath, inlineUrlDefs) = urlMapper(
+          val UrlMapResponse(pathDecl, pathTypes, maybeSecurityPath, inlineUrlDefs) = UrlEndpointComponent.urlMapper(
             p.url,
             name,
             targetScala3,
@@ -231,7 +190,7 @@ class EndpointGenerator {
           val SecurityDefn(securityDecl, securityTypes, securityWrappers) =
             SecurityGenerator.security(securitySchemes, m.security.getOrElse(doc.security))
           val (inParams, maybeLocalEnums, inTypes, inlineInDefns) =
-            ins(
+            InComponent.ins(
               m.resolvedParameters,
               m.requestBody.map(_.resolve(doc)),
               name,
@@ -248,7 +207,7 @@ class EndpointGenerator {
               seperateFilesForModels
             )
           val (outDecl, outTypes, errTypes, inlineDefns) =
-            outs(
+            OutComponent.outs(
               m.responses.map(_.resolve(doc)),
               streamingImplementation,
               doc,
@@ -312,7 +271,7 @@ class EndpointGenerator {
               case OpenapiSchemaArray(simple: OpenapiSchemaSimpleType, _, _, _)       =>
                 val name = mapSchemaSimpleTypeToType(simple)._1
                 s"List[$name]"
-              case simple: OpenapiSchemaSimpleType => mapSchemaSimpleTypeToType(simple)._1
+              case simple: OpenapiSchemaSimpleType                         => mapSchemaSimpleTypeToType(simple)._1
               case OpenapiSchemaMap(simple: OpenapiSchemaSimpleType, _, _) =>
                 val name = mapSchemaSimpleTypeToType(simple)._1
                 s"Map[String, $name]"
@@ -348,301 +307,6 @@ class EndpointGenerator {
     )
   }
 
-  private def urlMapper(
-      url: String,
-      endpointName: String,
-      targetScala3: Boolean,
-      jsonSerdeLib: JsonSerdeLib,
-      parameters: Seq[OpenapiParameter],
-      securityPathPrefixes: Seq[String],
-      doc: OpenapiDocument,
-      generateValidators: Boolean,
-      isReused: Boolean,
-      packageReuse: PackageReuseContext
-  )(implicit
-      location: Location
-  ): UrlMapResponse = {
-    val securityPrefixes = securityPathPrefixes.filter(url.startsWith)
-    // .in(("books" / path[String]("genre") / path[Int]("year")).mapTo[BooksFromYear])
-    val maxSecurityPrefix = if (securityPrefixes.nonEmpty) Some(securityPrefixes.maxBy(_.length)) else None
-    val inUrl = maxSecurityPrefix.fold(url)(url.stripPrefix)
-    def toPathDecl(url: String): (Array[String], Array[Option[String]], Array[Option[String]]) = url
-      .split('/')
-      .filter(_.nonEmpty)
-      .map { segment =>
-        if (segment.startsWith("{")) {
-          val name = segment.drop(1).dropRight(1)
-          val param = parameters.find(_.name == name)
-          param.fold(bail(s"URLParam $name not found!")) { p =>
-            p.schema match {
-              case st: OpenapiSchemaSimpleType =>
-                val (t, _) = mapSchemaSimpleTypeToType(st)
-                val desc = p.description.fold("")(d => s""".description("$d")""")
-                val validations = if (generateValidators) ValidationGenerator.mkValidations(doc, st, true) else ""
-                (s"""path[$t]("$name")$validations$desc""", Some(t), None)
-              case e: OpenapiSchemaEnum =>
-                val (param, inlineDefn, tpe, enumName) = getEnumParamDefn(endpointName, targetScala3, jsonSerdeLib, p, e, false)
-                val defns =
-                  if (isReused)
-                    Some(
-                      s"type $enumName = ${packageReuse.dependencyModelPath}.$enumName\nval $enumName = ${packageReuse.dependencyModelPath}.$enumName"
-                    )
-                  else inlineDefn.map(_.mkString("\n"))
-                (param, Some(tpe), defns)
-              case _ => bail("Can't create non-simple params to url yet")
-            }
-          }
-        } else {
-          ('"' + segment + '"', None, None)
-        }
-      }
-      .unzip3
-    val (inPath, tpes, inlineDefns) = toPathDecl(inUrl)
-    val inPathDecl = if (inPath.nonEmpty) ".in((" + inPath.mkString(" / ") + "))" else ""
-    val (secPathDecl, secInlineDefs) =
-      maxSecurityPrefix
-        .map(toPathDecl)
-        .map { case (ds, ts, inline) => (".prependSecurityIn(" + ds.mkString(" / ") + ")" -> ts.toSeq.flatten) -> inline }
-        .unzip
-    val flatInlineSecDefs: Seq[String] = secInlineDefs.toSeq.flatten.flatten
-    val flatInlineDefs: Seq[String] = inlineDefns.toSeq.flatten
-    UrlMapResponse(inPathDecl, tpes.toSeq.flatten, secPathDecl.headOption, flatInlineSecDefs ++ flatInlineDefs)
-  }
-
-  private def toOutType(baseType: String, isArray: Boolean, noOptionWrapper: Boolean) = (isArray, noOptionWrapper) match {
-    case (true, true)   => s"List[$baseType]"
-    case (true, false)  => s"Option[List[$baseType]]"
-    case (false, true)  => baseType
-    case (false, false) => s"Option[$baseType]"
-  }
-
-  private def getEnumParamDefn(
-      endpointName: String,
-      targetScala3: Boolean,
-      jsonSerdeLib: JsonSerdeLib,
-      param: OpenapiParameter,
-      e: OpenapiSchemaEnum,
-      isArray: Boolean
-  ): (String, Some[Seq[String]], String, String) = {
-    val enumName = endpointName.capitalize + strippedToCamelCase(param.name).capitalize
-    val enumParamRefs = if (param.in == "query" || param.in == "path") Set(enumName) else Set.empty[String]
-    val enumDefn = EnumGenerator.generateEnum(
-      enumName,
-      e,
-      targetScala3,
-      enumParamRefs,
-      jsonSerdeLib,
-      Set.empty
-    )
-
-    def arrayType = if (param.isExploded) "ExplodedValues" else "CommaSeparatedValues"
-
-    val tpe = if (isArray) s"$arrayType[$enumName]" else enumName
-    val required = param.required.getOrElse(false)
-    // 'exploded' params have no distinction between an empty list and an absent value, so don't wrap in 'Option' for them
-    val noOptionWrapper = required || (isArray && param.isExploded)
-    val req = if (noOptionWrapper) tpe else s"Option[$tpe]"
-    val outType = toOutType(enumName, isArray, noOptionWrapper)
-
-    def mapToList =
-      if (!isArray) "" else if (noOptionWrapper) s".map(_.values)($arrayType(_))" else s".map(_.map(_.values))(_.map($arrayType(_)))"
-
-    val desc = param.description.map(d => JavaEscape.escapeString(d)).fold("")(d => s""".description("$d")""")
-    (s"""${param.in}[$req]("${param.name}")$mapToList$desc""", Some(enumDefn), outType, enumName)
-  }
-
-  private def genParamDefn(
-      endpointName: String,
-      targetScala3: Boolean,
-      jsonSerdeLib: JsonSerdeLib,
-      param: OpenapiParameter,
-      doc: OpenapiDocument,
-      generateValidators: Boolean
-  )(implicit
-      location: Location
-  ): (String, Option[Seq[String]], String) =
-    param.schema match {
-      case st: OpenapiSchemaSimpleType =>
-        val (t, _) = mapSchemaSimpleTypeToType(st)
-        val required = param.required.getOrElse(false)
-        val req = if (required) t else s"Option[$t]"
-        val desc = param.description.map(JavaEscape.escapeString).fold("")(d => s""".description("$d")""")
-        val validation = if (generateValidators) ValidationGenerator.mkValidations(doc, st, required) else ""
-        (s"""${param.in}[$req]("${param.name}")$validation$desc""", None, req)
-      case OpenapiSchemaArray(st: OpenapiSchemaSimpleType, _, _, _) =>
-        val (t, _) = mapSchemaSimpleTypeToType(st)
-        val arrayType = if (param.isExploded) "ExplodedValues" else "CommaSeparatedValues"
-        val arr = s"$arrayType[$t]"
-        val required = param.required.getOrElse(false)
-        // 'exploded' params have no distinction between an empty list and an absent value, so don't wrap in 'Option' for them
-        val noOptionWrapper = required || param.isExploded
-        val req = if (noOptionWrapper) arr else s"Option[$arr]"
-
-        def mapToList = if (noOptionWrapper) s".map(_.values)($arrayType(_))" else s".map(_.map(_.values))(_.map($arrayType(_)))"
-
-        val desc = param.description.map(JavaEscape.escapeString).fold("")(d => s""".description("$d")""")
-        val outType = toOutType(t, true, noOptionWrapper)
-        (s"""${param.in}[$req]("${param.name}")$mapToList$desc""", None, outType)
-      case e @ OpenapiSchemaEnum(_, _, _) =>
-        getEnumParamDefn(endpointName, targetScala3, jsonSerdeLib, param, e, isArray = false) match {
-          case (a, b, c, _) => (a, b, c)
-        }
-      case OpenapiSchemaArray(e: OpenapiSchemaEnum, _, _, _) =>
-        getEnumParamDefn(endpointName, targetScala3, jsonSerdeLib, param, e, isArray = true) match {
-          case (a, b, c, _) => (a, b, c)
-        }
-      case x => bail(s"Can't create non-simple params - found $x")
-    }
-
-  private def aliases(packageReuse: PackageReuseContext, types: Seq[String], seperateFilesForModels: Boolean): String =
-    types.map(PackageReuseContext.enumAliasType(_, packageReuse, seperateFilesForModels)).mkString("\n")
-
-  private def ins(
-      parameters: Seq[OpenapiParameter],
-      requestBody: Option[OpenapiRequestBodyDefn],
-      endpointName: String,
-      targetScala3: Boolean,
-      jsonSerdeLib: JsonSerdeLib,
-      xmlSerdeLib: XmlSerdeLib,
-      streamingImplementation: StreamingImplementation,
-      doc: OpenapiDocument,
-      tapirCodegenDirectives: Set[String],
-      validators: ValidationDefns,
-      generateValidators: Boolean,
-      isReused: Boolean,
-      packageReuse: PackageReuseContext,
-      seperateFilesForModels: Boolean
-  )(implicit location: Location): (String, Option[String], Seq[String], Option[String]) = {
-
-    // .in(query[Limit]("limit").description("Maximum number of books to retrieve"))
-    // .in(header[AuthToken]("X-Auth-Token"))
-    val (params, maybeEnumDefns, inTypes) = parameters
-      .filter(_.in != "path")
-      .map { param =>
-        genParamDefn(endpointName, targetScala3, jsonSerdeLib, param, doc, generateValidators)
-      }
-      .map { case (defn, enums, tpe) => (s".in($defn)", enums, tpe) }
-      .unzip3
-
-    def mapContent(
-        content: OpenapiRequestBodyContent,
-        required: Boolean,
-        preferEager: Boolean = false
-    ): (String, String, Option[String]) = {
-      val schemaIsNullable = content.schema.nullable || (content.schema match {
-        case ref: OpenapiSchemaRef =>
-          doc.components.flatMap(_.schemas.get(ref.stripped).map(_.nullable)).contains(true)
-        case _ => false
-      })
-      val MappedContentType(decl, tpe, maybeInlineDefn, inlineTypes) =
-        contentTypeMapper(
-          content.contentType,
-          content.schema,
-          streamingImplementation,
-          required && !schemaIsNullable,
-          endpointName,
-          Request,
-          preferEager,
-          xmlSerdeLib,
-          tapirCodegenDirectives,
-          validators
-        )
-
-      val inlineDefn = maybeInlineDefn.map(d => if (isReused) aliases(packageReuse, inlineTypes, seperateFilesForModels) else d)
-      (decl, tpe, inlineDefn)
-    }
-    val (rqBody, maybeReqType, maybeInlineDefns) = requestBody.flatMap { b =>
-      if (b.content.isEmpty) None
-      else if (b.content.size == 1) {
-        val content: OpenapiModels.OpenapiRequestBodyContent = b.content.head
-        val (decl, tpe, maybeInlineDefn) = mapContent(content, b.required)
-        val d = b.description.map(s => s""".description("${JavaEscape.escapeString(s)}")""").getOrElse("")
-        Some((s".in($decl$d)", tpe, maybeInlineDefn))
-      } else {
-        // We cannot mix eager and streaming types when using oneOfBody
-        val preferEager = b.content.exists(c => eagerTypes.contains(c.contentType))
-        val mapped = b.content.map(mapContent(_, b.required, preferEager))
-        val (decls, tpes, maybeInlineDefns) = mapped.unzip3
-        val distinctTypes = tpes.distinct
-        // If the types are distinct, we need to produce wrappers with a common parent for oneOfBody to work. If they're
-        // eager or lazy binary, wrappers make it easier to implement logic binding.
-        val needsAliases =
-          distinctTypes.size != 1 || tpes.head == "Array[Byte]" || tpes.head.contains("BinaryStream") || tpes.head.contains("fs2.Stream")
-        val tpesAreBin = tpes.head.contains("BinaryStream") || tpes.head.contains("fs2.Stream")
-        def wrapBinType(s: String) = if (tpesAreBin) s"sttp.tapir.EndpointIO.StreamBodyWrapper($s)" else s
-        val traitName = s"${endpointName.capitalize}BodyIn"
-        val declsByWrapperClassName = decls
-          .zip(tpes)
-          .zipWithIndex
-          .map { case ((decl, t), i) =>
-            val caseClassName =
-              if (t == "Array[Byte]" || t.contains("BinaryStream") || tpes.head.contains("fs2.Stream"))
-                s"${endpointName.capitalize}Body${i}In"
-              else s"${endpointName.capitalize}Body${t.split('.').last.replaceAll("[\\]\\[]", "_")}In"
-            (caseClassName, t, decl)
-          }
-          .groupBy(_._1)
-        val aliasDefns =
-          if (needsAliases && isReused) {
-            val wrappers = declsByWrapperClassName
-              .map { case (name, _) =>
-                s"type $name = ${packageReuse.dependencyModelPath}.$name\nval $name = ${packageReuse.dependencyModelPath}.$name\n"
-              }
-              .toSeq
-              .sorted
-              .mkString("\n")
-            Some(s"""
-                    |type $traitName = ${packageReuse.dependencyModelPath}.$traitName
-                    |$wrappers
-                    |""".stripMargin)
-          } else if (needsAliases) {
-            val wrappers = declsByWrapperClassName
-              .map { case (name, seq) => s"""case class ${name}(value: ${seq.head._2}) extends $traitName""" }
-              .toSeq
-              .sorted
-              .mkString("\n")
-            Some(s"""
-               |sealed trait $traitName extends Product with java.io.Serializable
-               |$wrappers
-               |""".stripMargin)
-          } else None
-        val classNameByDecl = declsByWrapperClassName.flatMap { case (className, seq) =>
-          seq.map { case (_, _, decl) => decl -> className }
-        }
-
-        val tpe = if (needsAliases) traitName else tpes.head
-        val distinctInlineDefns = maybeInlineDefns.flatten.distinct.mkString("\n")
-        val didO = if (distinctInlineDefns.isEmpty) None else Some(distinctInlineDefns)
-        val d = b.description.map(s => s""".description("${JavaEscape.escapeString(s)}")""").getOrElse("")
-        val bodies =
-          if (needsAliases)
-            decls.zip(tpes).map { case (decl, _) =>
-              wrapBinType(s"$decl.map(${classNameByDecl(decl)}(_))(_.value).widenBody[$traitName]$d")
-            }
-          else decls.map(_ + d)
-        Some(
-          (
-            s""".in(oneOfBody[$tpe](
-             |${indent(2)(bodies.mkString(",\n"))}))""".stripMargin,
-            tpe,
-            combine(didO.filterNot(_ => isReused), aliasDefns)
-          )
-        )
-      }
-    }.unzip3
-
-    (
-      (params ++ rqBody).mkString("\n"),
-      maybeEnumDefns.foldLeft(Option.empty[String]) {
-        case (acc, None)            => acc
-        case (None, Some(nxt))      => Some(nxt.mkString("\n"))
-        case (Some(acc), Some(nxt)) => Some(acc + "\n" + nxt.mkString("\n"))
-      },
-      inTypes ++ maybeReqType,
-      maybeInlineDefns.foldLeft(Option.empty[String])(combine(_, _))
-    )
-  }
-
   private def tags(openapiTags: Option[Seq[String]]): String = {
     // .tags(List("A", "B"))
     openapiTags.map(_.distinct.mkString(".tags(List(\"", "\", \"", "\"))")).mkString
@@ -660,612 +324,6 @@ class EndpointGenerator {
       .mkString("\n")
   }
   else None
-
-  // treats redirects as ok
-  private val okStatus = """([23]\d\d)""".r
-  private val errorStatus = """([45]\d\d)""".r
-  private def outs(
-      responses: Seq[OpenapiResponseDef],
-      streamingImplementation: StreamingImplementation,
-      doc: OpenapiDocument,
-      targetScala3: Boolean,
-      endpointName: String,
-      jsonSerdeLib: JsonSerdeLib,
-      xmlSerdeLib: XmlSerdeLib,
-      tapirCodegenDirectives: Set[String],
-      validators: ValidationDefns,
-      generateValidators: Boolean,
-      isReused: Boolean,
-      packageReuse: PackageReuseContext,
-      seperateFilesForModels: Boolean
-  )(implicit
-      location: Location
-  ) = {
-    // .errorOut(stringBody)
-    // .out(oneOfBody(jsonBody[List[Book]]))
-
-    val (outs, errorOuts) = responses.partition { resp =>
-      resp.code match {
-        case okStatus(_)                => true
-        case "default" | errorStatus(_) => false
-        case x                          => bail(s"Statuscode mapping is incomplete! Cannot handle $x")
-      }
-    }
-
-    def bodyFmt(resp: OpenapiResponseDef, isErrorPosition: Boolean, optional: Boolean = false): (String, Option[String], Option[String]) = {
-      def wrapContent(content: OpenapiResponseContent, preferEager: Boolean = false) = {
-        val schemaIsNullable = content.schema.nullable || (content.schema match {
-          case ref: OpenapiSchemaRef =>
-            doc.components.flatMap(_.schemas.get(ref.stripped).map(_.nullable)).contains(true)
-          case _ => false
-        })
-        val MappedContentType(decl, tpe, maybeInlineDefn, inlineTypes) = {
-          contentTypeMapper(
-            content.contentType,
-            content.schema,
-            streamingImplementation,
-            !(optional || schemaIsNullable),
-            endpointName,
-            if (isErrorPosition) Err else Response,
-            preferEager,
-            xmlSerdeLib,
-            tapirCodegenDirectives,
-            validators
-          )
-        }
-        val inlineDefn = maybeInlineDefn.map(d => if (isReused) aliases(packageReuse, inlineTypes, seperateFilesForModels) else d)
-        (decl, tpe, inlineDefn)
-      }
-      resp.content match {
-        case Nil            => ("", None, None)
-        case content +: Nil =>
-          val (decl, tpe, maybeInlineDefn) = wrapContent(content)
-          val d = s""".description("${JavaEscape.escapeString(resp.description)}")"""
-          (s"$decl$d", Some(tpe), maybeInlineDefn)
-        case seq =>
-          // We cannot mix eager and streaming types when using oneOfBody
-          val preferEager = seq.exists(c => eagerTypes.contains(c.contentType))
-          val (decls, tpes, maybeInlineDefns) = seq.map(wrapContent(_, preferEager)).unzip3
-          val distinctTypes = tpes.distinct
-          // If the types are distinct, we need to produce wrappers with a common parent for oneOfBody to work. If they're
-          // eager or lazy binary, wrappers make it easier to implement logic binding.
-          val needsAliases =
-            distinctTypes.size != 1 || tpes.head == "Array[Byte]" || tpes.head.contains("BinaryStream") || tpes.head.contains("fs2.Stream")
-          val tpesAreBin = tpes.head.contains("BinaryStream") || tpes.head.contains("fs2.Stream")
-          def wrapBinType(s: String) = if (tpesAreBin) s"sttp.tapir.EndpointIO.StreamBodyWrapper($s)" else s
-          val suff = if (isErrorPosition) "Err" else "Out"
-          val traitName = s"${endpointName.capitalize}Body$suff"
-          val d = s""".description("${JavaEscape.escapeString(resp.description)}")"""
-          val declsByWrapperClassName = decls
-            .zip(tpes)
-            .zip(seq.map(_.contentType))
-            .zipWithIndex
-            .map { case (((decl, t), ct), i) =>
-              val caseClassName =
-                if (t == "Array[Byte]" || t.contains("BinaryStream") || tpes.head.contains("fs2.Stream"))
-                  s"${endpointName.capitalize}Body${i}$suff"
-                else s"${endpointName.capitalize}Body${t.split('.').last.replaceAll("[\\]\\[]", "_")}$suff"
-              (caseClassName, t, decl, ct)
-            }
-            .groupBy(_._1)
-          val aliasDefns =
-            if (needsAliases && isReused) {
-              val parentModelPath = packageReuse.modelRoot(seperateFilesForModels)
-              val wrappers = declsByWrapperClassName
-                .map { case (name, seq) =>
-                  s"type $name = $parentModelPath.$name\nval $name = $parentModelPath.$name\n"
-                }
-                .toSeq
-                .sorted
-                .mkString("\n")
-              Some(s"""
-                      |type $traitName = $parentModelPath.$traitName
-                      |type ${traitName}Full = $parentModelPath.${traitName}Full
-                      |val ${traitName}Full = $parentModelPath.${traitName}Full
-                      |$wrappers
-                      |""".stripMargin)
-            } else if (needsAliases) {
-              def callers(tpe: String, impl: Boolean) = declsByWrapperClassName
-                .flatMap { case (_, seq) =>
-                  seq.map { case (_, t, _, ct) =>
-                    s"""$tpe `$ct`: () => $t${if (impl) s""" = () => throw new RuntimeException("Body for content type $ct not provided")"""
-                      else ","}"""
-                  }
-                }
-                .toSeq
-                .sorted
-                .mkString("\n")
-              val wrappers = declsByWrapperClassName
-                .map { case (name, seq) =>
-                  val defns = seq.map { case (_, t, _, ct) => s"""override def `$ct`: () => $t = () => value""" }.sorted.mkString("\n")
-                  s"""case class ${name}(value: ${seq.head._2}) extends $traitName{
-                     |${indent(2)(defns)}
-                     |}""".stripMargin
-                }
-                .toSeq
-                .sorted
-                .mkString("\n")
-              Some(s"""
-                      |sealed trait $traitName extends Product with java.io.Serializable {
-                      |${indent(2)(callers("def", true))}
-                      |}
-                      |case class ${traitName}Full (
-                      |${indent(2)(callers("override val", false))}
-                      |) extends $traitName
-                      |$wrappers
-                      |""".stripMargin)
-            } else None
-          val classNameByDecl = declsByWrapperClassName.flatMap { case (className, seq) =>
-            seq.map { case (_, _, decl, _) => decl -> className }
-          }
-
-          val tpe = if (needsAliases) traitName else tpes.head
-          val bodies =
-            if (needsAliases)
-              decls.zip(tpes).zip(seq.map(_.contentType)).map { case ((decl, _), ct) =>
-                wrapBinType(
-                  s"$decl.map(${classNameByDecl(decl)}(_))(_.`$ct`())\n" +
-                    s".map(_.asInstanceOf[$traitName])(p => ${classNameByDecl(decl)}(p.`$ct`()))$d"
-                )
-              }
-            else decls.map(_ + d)
-          val distinctInlineDefns = maybeInlineDefns.flatten.distinct.mkString("\n")
-          val didO = if (distinctInlineDefns.isEmpty) None else Some(distinctInlineDefns)
-          (
-            s"""oneOfBody[$tpe](
-               |${indent(2)(bodies.mkString(",\n"))})""".stripMargin,
-            Some(tpe),
-            combine(didO.filterNot(_ => isReused), aliasDefns)
-          )
-      }
-    }
-    def mappedGroup(group: Seq[OpenapiResponseDef], isErrorPosition: Boolean): MappedOutGroup =
-      group match {
-        case Nil         => MappedOutGroup(None, None, None)
-        case resp +: Nil =>
-          val (outHeaderDefns, outHeaderInlineEnums, outHeaderTypes) = resp.getHeaders.map { case (name, defn) =>
-            genParamDefn(endpointName, targetScala3, jsonSerdeLib, defn.resolved(name, doc).param, doc, generateValidators)
-          }.unzip3
-          val hs = outHeaderDefns.map(d => s".and($d)").mkString
-          def ht(wrap: Boolean = true) =
-            if (outHeaderTypes.isEmpty) None
-            else if (outHeaderTypes.size == 1) Some(outHeaderTypes.head)
-            else if (!wrap) Some(outHeaderTypes.mkString(", "))
-            else Some(s"(${outHeaderTypes.mkString(", ")})")
-          def inlineHeaderEnumDefns = outHeaderInlineEnums.foldLeft(Seq.empty[String]) { (acc, next) => acc ++ next.toSeq.flatten } match {
-            case Nil => None
-            case s   => Some(s.mkString("\n"))
-          }
-          resp.content match {
-            case Nil =>
-              val d = s""".description("${JavaEscape.escapeString(resp.description)}")"""
-              MappedOutGroup(
-                resp.code match {
-                  case "200" | "default" if outHeaderDefns.isEmpty => None
-                  case "200"                                       => Some(s"statusCode(sttp.model.StatusCode(200))$d$hs")
-                  case "default"                                   => Some(s"statusCode(sttp.model.StatusCode(400))$d$hs")
-                  case okStatus(s)                                 => Some(s"statusCode(sttp.model.StatusCode($s))$d$hs")
-                  case errorStatus(s)                              => Some(s"statusCode(sttp.model.StatusCode($s))$d$hs")
-                },
-                ht(),
-                inlineHeaderEnumDefns
-              )
-            case _ =>
-              val (decl, maybeBodyType, inlineDefn) = bodyFmt(resp, isErrorPosition)
-              val tpe =
-                if (outHeaderTypes.isEmpty) maybeBodyType
-                else if (maybeBodyType.isEmpty) ht()
-                else maybeBodyType.map(t => s"($t, ${ht(false).get})")
-              val tpeIsBin = maybeBodyType.exists(t => t.contains("BinaryStream") || t.contains("fs2.Stream"))
-              MappedOutGroup(
-                Some(resp.code match {
-                  case "200" | "default" if !tpeIsBin || hs.isEmpty => s"$decl$hs"
-                  case "200" | "default"                            => s"$decl.toEndpointIO$hs"
-                  case okStatus(s) if tpeIsBin                      => s"$decl.toEndpointIO$hs.and(statusCode(sttp.model.StatusCode($s)))"
-                  case okStatus(s)                                  => s"$decl$hs.and(statusCode(sttp.model.StatusCode($s)))"
-                  case errorStatus(s)                               => s"$decl$hs.and(statusCode(sttp.model.StatusCode($s)))"
-                }),
-                tpe,
-                inlineDefn.map(_ ++ inlineHeaderEnumDefns.getOrElse("")).orElse(inlineHeaderEnumDefns)
-              )
-          }
-        case many =>
-          if (many.map(_.code).distinct.size != many.size) bail("Cannot construct schema for multiple responses with same status code")
-          val contentCanBeEmpty = many.exists(_.content.isEmpty)
-          val allBodiesAreEmpty = many.forall(_.content.isEmpty)
-          val allResponsesAreEmpty = allBodiesAreEmpty && many.forall(_.getHeaders.isEmpty)
-          val (noHeaders, hs, outHeaderDefns, matchHeaders, headerTypes, headerTopType) =
-            headerDefns(targetScala3, jsonSerdeLib, doc, generateValidators)(endpointName, many)
-          val (oneOfs, types, inlineDefns) = many.map { m =>
-            val (decl, maybeBodyType, inlineDefn1) = bodyFmt(m, isErrorPosition, optional = contentCanBeEmpty)
-            val code = if (m.code == "default") "400" else m.code
-            if (decl == "" && allResponsesAreEmpty && noHeaders)
-              (
-                s"oneOfVariantSingletonMatcher(sttp.model.StatusCode($code), " +
-                  s"""emptyOutput.description("${JavaEscape.escapeString(m.description)}"))(())""",
-                maybeBodyType,
-                inlineDefn1
-              )
-            else if (decl == "" && noHeaders)
-              (
-                s"oneOfVariantSingletonMatcher(sttp.model.StatusCode($code), " +
-                  s"""emptyOutput.description("${JavaEscape.escapeString(m.description)}"))(None)""",
-                maybeBodyType,
-                inlineDefn1
-              )
-            else if (decl == "") {
-              val s =
-                if (allBodiesAreEmpty)
-                  s"oneOfVariantValueMatcher(sttp.model.StatusCode($code), " +
-                    s"""emptyOutput.description("${JavaEscape.escapeString(m.description)}")""" +
-                    s""".and(${hs(m)})){ case ${matchHeaders(m)} => true}"""
-                else
-                  s"oneOfVariantValueMatcher(sttp.model.StatusCode($code), " +
-                    s"""emptyOutputAs(None).description("${JavaEscape.escapeString(
-                        m.description
-                      )}").and(${hs(m)})){ case (None, ${matchHeaders(m)}) => true}"""
-              (s, maybeBodyType, inlineDefn1)
-            } else {
-              def bodyAndHeaderTypes(bodyType: String): String =
-                if (noHeaders) bodyType
-                else if (allBodiesAreEmpty) headerTypes(m)
-                else s"($bodyType, ${headerTypes(m)})"
-              def matchBodyAndHeaders(matchBody: String): String = if (noHeaders) matchBody else s"($matchBody, ${matchHeaders(m)})"
-              val tpeIsBin = maybeBodyType.exists(t => t.contains("BinaryStream") || t.contains("fs2.Stream"))
-              val maybeStrict = if (tpeIsBin) ".toEndpointIO" else ""
-              def h = hs(m) match { case s if s.isEmpty => ""; case s => s".and($s)" }
-              if (contentCanBeEmpty) {
-                val (_, nonOptionalType, _) = bodyFmt(m, isErrorPosition)
-                val maybeMap = if (m.content.size > 1 || tpeIsBin) ".map(Some(_))(_.orNull)" else ""
-                val someType = nonOptionalType.map(": " + _.replaceAll("^Option\\[(.+)]$", "$1")).getOrElse("")
-                (
-                  s"oneOfVariantValueMatcher(sttp.model.StatusCode(${code}), $decl$maybeStrict$maybeMap$h){ case ${matchBodyAndHeaders(s"Some(_$someType)")} => true }",
-                  maybeBodyType,
-                  inlineDefn1
-                )
-              } else
-                (
-                  s"oneOfVariant${maybeBodyType.map(s => s"[${bodyAndHeaderTypes(s)}]").getOrElse("")}(sttp.model.StatusCode(${code}), $decl$maybeStrict$h)",
-                  maybeBodyType,
-                  inlineDefn1
-                )
-            }
-          }.unzip3
-          val parentMap = doc.components.toSeq
-            .flatMap(_.schemas)
-            .collect { case (k, v: OpenapiSchemaOneOf) =>
-              v.types.map {
-                case r: OpenapiSchemaRef        => r.stripped -> k
-                case x: OpenapiSchemaSimpleType => mapSchemaSimpleTypeToType(x)._1 -> k
-                case x                          => bail(s"Unexpected oneOf child type $x")
-              }
-            }
-            .flatten
-            .groupBy(_._1)
-            .map { case (k, vs) => k -> vs.map(_._2) }
-            .toMap
-          val traitName = s"${endpointName.capitalize}Body${if (isErrorPosition) "Err" else "Out"}"
-          val mappable = Set("application/json", "application/xml", "multipart/form-data")
-          val bodyIsStreaming = (!isErrorPosition && tapirCodegenDirectives.contains(forceRespStreaming)) ||
-            (!isErrorPosition && tapirCodegenDirectives.contains(forceStreaming))
-          val bodyIsEager = !bodyIsStreaming && (isErrorPosition ||
-            (!isErrorPosition && tapirCodegenDirectives.contains(forceRespEager)) ||
-            (!isErrorPosition && tapirCodegenDirectives.contains(forceEager)))
-          val allElemTypes = many
-            .flatMap(y =>
-              y.content.map(x =>
-                (x.contentType, x.schema, y.content.size > 1 && y.content.map(_.contentType).exists(!mappable.contains(_)))
-              )
-            )
-            .map {
-              case (_, _, _) if bodyIsStreaming                            => capabilityType(streamingImplementation)
-              case (_, _, true)                                            => traitName
-              case (ct, _, _) if ct.startsWith("text/") && isErrorPosition => "String"
-              case ("text/plain" | "text/html", _, _)                      => "String"
-              case ("application/json", _, _) if tapirCodegenDirectives.contains(jsonBodyAsString) => "String"
-              case (ct, r: OpenapiSchemaRef, _) if mappable.contains(ct)                           => r.stripped
-              case (ct, x: OpenapiSchemaSimpleType, _) if mappable.contains(ct)                    => mapSchemaSimpleTypeToType(x)._1
-              case (ct, x, _) if mappable.contains(ct) => bail(s"Unexpected oneOf elem type $x with content type $ct")
-              case (_, _, _) if bodyIsEager            => "Array[Byte]"
-              case (_, _, _)                           => capabilityType(streamingImplementation)
-            }
-            .distinct
-          val commmonType = {
-            if (allResponsesAreEmpty || allElemTypes.isEmpty) "Unit"
-            else if (contentCanBeEmpty && allElemTypes.size == 1) s"Option[${allElemTypes.head}]"
-            else if (allElemTypes.size == 1) allElemTypes.head
-            else {
-              val baseType = allElemTypes.map { s => parentMap.getOrElse(s, Nil).toSet }.reduce(_ intersect _) match {
-                case s if s.isEmpty && targetScala3 => types.flatten.mkString(" | ")
-                case s if s.isEmpty                 => "Any"
-                case s if targetScala3              => s.mkString(" & ")
-                case s                              => s.mkString(" with ")
-              }
-              if (contentCanBeEmpty) s"Option[$baseType]" else baseType
-            }
-          }
-          val oneOfType = if (noHeaders) commmonType else if (allBodiesAreEmpty) headerTopType else s"($commmonType, $headerTopType)"
-          MappedOutGroup(
-            Some(s"oneOf[$oneOfType](${oneOfs.mkString("\n  ", ",\n  ", "")})"),
-            Some(oneOfType),
-            (inlineDefns ++ outHeaderDefns).foldLeft(Option.empty[String])(combine(_, _))
-          )
-      }
-
-    val MappedOutGroup(outDecls, outTypes, inlineOutDefns) = mappedGroup(outs, false)
-    val mappedOuts = outDecls.map(s => s".out($s)")
-    val MappedOutGroup(errDecls, errTypes, inlineErrDefns) = mappedGroup(errorOuts, true)
-    val mappedErrorOuts = errDecls.map(s => s".errorOut($s)")
-
-    (Seq(mappedErrorOuts, mappedOuts).flatten.mkString("\n"), outTypes, errTypes, combine(inlineOutDefns, inlineErrDefns))
-  }
-
-  private def headerDefns(targetScala3: Boolean, jsonSerdeLib: JsonSerdeLib, doc: OpenapiDocument, generateValidators: Boolean)(
-      endpointName: String,
-      many: Seq[OpenapiResponseDef]
-  )(implicit
-      location: Location
-  ): (Boolean, OpenapiResponseDef => String, Seq[Option[String]], OpenapiResponseDef => String, OpenapiResponseDef => String, String) = {
-    val (paramNames, headerNamesAndTypes) = many.map { m =>
-      m.getHeaders
-        .map { case (name, defn) =>
-          val param = defn.resolved(name, doc).param
-          NameHelpers.safeVariableName(param.name) ->
-            genParamDefn(endpointName, targetScala3, jsonSerdeLib, param, doc, generateValidators)
-        }
-        .toSeq
-        .sortBy(_._1)
-        .unzip
-    }.unzip
-    if (headerNamesAndTypes.forall(_.isEmpty)) (true, _ => "", Nil, _ => "", _ => "", "")
-    else if (headerNamesAndTypes.map(_.map { case (name, _, defn) => name -> defn }.toSet).distinct.size == 1) {
-      val commonResponseHeaders = headerNamesAndTypes.head
-      val (outHeaderDefns, outHeaderInlineEnums, outHeaderTypes) = commonResponseHeaders.unzip3
-      val underscores = outHeaderDefns.map(_ => "_").mkString(", ")
-      val hs = (_: OpenapiResponseDef) => outHeaderDefns.zipWithIndex.map { case (d, 0) => d; case (d, _) => s".and($d)" }.mkString
-      val noHeaders = commonResponseHeaders.isEmpty
-
-      val ht = (_: OpenapiResponseDef) =>
-        if (outHeaderTypes.isEmpty) bail("Should not try to construct header types if no headers are required")
-        else if (outHeaderTypes.size == 1) outHeaderTypes.head
-        else s"(${outHeaderTypes.mkString(", ")})"
-
-      val headerTopType = if (outHeaderTypes.isEmpty) "Unit" else ht(null)
-
-      val enumDefns = outHeaderInlineEnums.map(_.map(_.mkString("\n")))
-      (noHeaders, hs, enumDefns, (_: OpenapiResponseDef) => underscores, ht, headerTopType)
-    } else {
-      val traitName = s"${endpointName.capitalize}ResponseHeader"
-
-      val (headerMappingDefns, headerModelDefns, enumDefns: Seq[Seq[String]]) = headerNamesAndTypes
-        .zip(many.map(_.code))
-        .zip(paramNames)
-        .map {
-          case ((s, c), _) if s.isEmpty =>
-            val objName = s"$traitName$c"
-            (s"emptyOutputAs($objName)", s"case object $objName extends $traitName", Seq.empty[String])
-          case ((s, c), names) =>
-            val className = s"$traitName$c"
-            val (headerDefns, headerInlineEnums, headerTypes) = s.unzip3
-            val rawOutput = headerDefns.zipWithIndex.map { case (d, 0) => d; case (d, _) => s".and($d)" }.mkString
-            val output =
-              if (s.size == 1) s"$rawOutput.map($className(_))(_.${names.head})"
-              else s"($rawOutput).map(($className.apply _).tupled)($className.unapply(_).get)"
-            val fields = names.zip(headerTypes).map { case (n, t) => s"$n: $t" }.mkString(", ")
-            (output, s"""case class $traitName$c($fields) extends $traitName""", headerInlineEnums.flatten.flatten)
-        }
-        .unzip3
-      val mappingsByCode = headerMappingDefns
-        .zip(many.map(_.code))
-        .map { case (s, c) => c -> s }
-        .toMap
-      val tpesByCode = headerModelDefns
-        .zip(many.map(_.code))
-        .map { case (s, c) => c -> s }
-        .toMap
-      val headerTypeDefns =
-        s"""sealed trait ${endpointName.capitalize}ResponseHeader
-           |${tpesByCode.values.toSeq.sorted.mkString("\n")}
-           |""".stripMargin
-      def ht(m: OpenapiResponseDef) = tpesByCode(m.code)
-      def getMapping(m: OpenapiResponseDef) = mappingsByCode(m.code)
-      def getMatch(m: OpenapiResponseDef) =
-        if (m.getHeaders.isEmpty) s"$traitName${m.code}"
-        else s"(_: $traitName${m.code})"
-      val enums = enumDefns.flatten.distinct.map(Some(_))
-      (false, getMapping, Some(headerTypeDefns) +: enums, getMatch, ht, traitName)
-    }
-  }
-
-  private def inlineDefn(endpointName: String, position: Position, schemaRef: OpenapiSchemaObject) = {
-    require(schemaRef.properties.forall(_._2.`type`.isInstanceOf[OpenapiSchemaSimpleType]))
-    val inlineClassName = endpointName.capitalize + position
-    val properties = schemaRef.properties.map { case (k, v) =>
-      val (st, nb) = mapSchemaSimpleTypeToType(v.`type`.asInstanceOf[OpenapiSchemaSimpleType], multipartForm = true)
-      val optional = !schemaRef.required.contains(k) || nb
-      val t = if (optional) s"Option[$st]" else st
-      val default = v.default
-        .map(j => " = " + DefaultValueRenderer.render(Map.empty, v.`type`, optional, RenderConfig())(j))
-        .getOrElse(if (optional) " = None" else "")
-      s"$k: $t$default"
-    }
-    val inlineClassDefn =
-      s"""case class $inlineClassName (
-         |${indent(2)(properties.mkString(",\n"))}
-         |)""".stripMargin
-    inlineClassName -> Some(inlineClassDefn)
-  }
-  private def contentTypeMapper(
-      contentType: String,
-      schema: OpenapiSchemaType,
-      streamingImplementation: StreamingImplementation,
-      required: Boolean,
-      endpointName: String,
-      position: Position,
-      defaultEager: Boolean,
-      xmlSerdeLib: XmlSerdeLib,
-      tapirCodegenDirectives: Set[String],
-      validators: ValidationDefns
-  )(implicit location: Location): MappedContentType = {
-    def vRef(t: OpenapiSchemaType, r: Boolean) = {
-      val maybeValidatorRef = t match {
-        case r: OpenapiSchemaRef => Some(r.stripped)
-        // TODO: Validation on inline defns
-        case _ => None
-      }
-      maybeValidatorRef
-        .flatMap(validators.defns.get)
-        .map(_.name)
-        .map {
-          case n if !r => s".validateOption(${n}Validator)"
-          case n       => s".validate(${n}Validator)"
-        }
-        .getOrElse("")
-    }
-    def v(r: Boolean) = vRef(schema, r)
-    val streaming = (tapirCodegenDirectives.contains(forceStreaming) && position != Err) ||
-      (tapirCodegenDirectives.contains(forceReqStreaming) && position == Request) ||
-      (tapirCodegenDirectives.contains(forceRespStreaming) && position == Response)
-    val eager = !streaming && (defaultEager ||
-      position == Err ||
-      (tapirCodegenDirectives.contains(forceEager) && position != Err) ||
-      (tapirCodegenDirectives.contains(forceReqEager) && position == Request) ||
-      (tapirCodegenDirectives.contains(forceRespEager) && position == Response))
-    contentType match {
-      case any if streaming =>
-        failoverBinaryCase(endpointName, position, any, schema, false, streamingImplementation)
-      case "text/plain" =>
-        MappedContentType("stringBody", "String")
-      case "text/html" =>
-        MappedContentType("htmlBodyUtf8", "String")
-      case "application/xml" if xmlSerdeLib != XmlSerdeLib.NoSupport =>
-        val (outT: String, maybeInline: Option[String], maybeAlias: Option[String], maybeTpe: Seq[String]) = schema match {
-          case st: OpenapiSchemaSimpleType =>
-            val (t, _) = mapSchemaSimpleTypeToType(st)
-            (t, None, None, Nil)
-          case a @ OpenapiSchemaArray(st: OpenapiSchemaSimpleType, _, _, _) =>
-            val (t, _) = mapSchemaSimpleTypeToType(st)
-            val tpeName = endpointName.capitalize + position
-            val xmlSeqConfig = XmlSerdeGenerator.genTopLevelSeqSerdes(xmlSerdeLib, a, endpointName, position.toString).map(_._2)
-            (s"List[$t]", xmlSeqConfig, Some(tpeName), xmlSeqConfig.toSeq.map(_ => tpeName))
-          case x => bail(s"Only ref, primitive (and arrays of either) schemas supported for xml body (found $x)")
-        }
-        val req = if (required) outT else s"Option[$outT]"
-        def toList = if (required) ".toList" else ".map(_.toList)"
-        val bodyType = maybeAlias.map(a => s"xmlBody[$a].map(_.asInstanceOf[$req]$toList)(_.asInstanceOf[$a])").getOrElse(s"xmlBody[$req]")
-        MappedContentType(bodyType + v(required), req, maybeInline, maybeTpe)
-      case "application/json" if tapirCodegenDirectives.contains(jsonBodyAsString) =>
-        if (required) MappedContentType("stringJsonBody", "String", None)
-        else MappedContentType("stringJsonBody.map(Option(_))(_.orNull)", "Option[String]", None)
-      case "application/json" =>
-        val (outT, maybeInline) = schema match {
-          case st: OpenapiSchemaSimpleType =>
-            val (t, _) = mapSchemaSimpleTypeToType(st)
-            t -> None
-          case OpenapiSchemaArray(st: OpenapiSchemaSimpleType, _, _, _) =>
-            val (t, _) = mapSchemaSimpleTypeToType(st)
-            s"List[$t]" -> None
-          case OpenapiSchemaMap(st: OpenapiSchemaSimpleType, _, _) =>
-            val (t, _) = mapSchemaSimpleTypeToType(st)
-            s"Map[String, $t]" -> None
-          case schemaRef: OpenapiSchemaObject if schemaRef.properties.forall(_._2.`type`.isInstanceOf[OpenapiSchemaSimpleType]) =>
-            inlineDefn(endpointName, position, schemaRef)
-          case x => bail(s"Can't create non-simple or array params as output (found $x)")
-        }
-        val req = if (required) outT else s"Option[$outT]"
-        MappedContentType(s"jsonBody[$req]" + v(required), req, maybeInline, maybeInline.map(_ => outT).toSeq)
-
-      case "multipart/form-data" =>
-        schema match {
-          case _: OpenapiSchemaBinary =>
-            MappedContentType("multipartBody", "Seq[Part[Array[Byte]]]")
-          case schemaRef: OpenapiSchemaRef =>
-            val (t, _) = mapSchemaSimpleTypeToType(schemaRef, multipartForm = true)
-            MappedContentType(s"multipartBody[$t]" + v(required), t)
-          case schemaRef: OpenapiSchemaObject if schemaRef.properties.forall(_._2.`type`.isInstanceOf[OpenapiSchemaSimpleType]) =>
-            val (inlineClassName, inlineClassDefn) = inlineDefn(endpointName, position, schemaRef)
-            MappedContentType(
-              s"multipartBody[$inlineClassName]" + v(required),
-              inlineClassName,
-              inlineClassDefn,
-              inlineClassDefn.map(_ => inlineClassName).toSeq
-            )
-          case x => bail(s"$contentType only supports schema ref or binary, or simple inline property maps with string values. Found $x")
-        }
-      case other =>
-        failoverBinaryCase(
-          endpointName,
-          position,
-          other,
-          schema,
-          eager,
-          streamingImplementation
-        )
-    }
-  }
-  private def failoverBinaryCase(
-      endpointName: String,
-      position: Position,
-      contentType: String,
-      schema: OpenapiSchemaType,
-      isEager: Boolean,
-      streamingImplementation: StreamingImplementation
-  )(implicit location: Location): MappedContentType = {
-    def codec(baseType: String, contentType: String) = baseType match {
-      case "Array[Byte]" =>
-        s"Codec.id[$baseType, `${contentType}CodecFormat`](`${contentType}CodecFormat`(), Schema.schemaForByteArray)"
-      case "String" =>
-        s"Codec.id[$baseType, `${contentType}CodecFormat`](`${contentType}CodecFormat`(), Schema.schemaForString)"
-    }
-
-    def eagerBody = contentType match {
-      case "application/octet-stream" => "rawBinaryBody(sttp.tapir.RawBodyType.ByteArrayBody)"
-      case o if o.startsWith("text/") => s"stringBodyUtf8AnyFormat(${codec("String", o)})"
-      case "application/xml"          => s"EndpointIO.Body(RawBodyType.ByteArrayBody, CodecFormat.Xml(), EndpointIO.Info.empty)"
-      case o                          => s"EndpointIO.Body(RawBodyType.ByteArrayBody, ${codec("Array[Byte]", o)}, EndpointIO.Info.empty)"
-    }
-    def streamingBody = contentType match {
-      case "text/plain"                        => "CodecFormat.TextPlain()"
-      case "text/html"                         => "CodecFormat.TextHtml()"
-      case "multipart/form-data"               => "CodecFormat.MultipartFormData()"
-      case "application/grpc"                  => "CodecFormat.Grpc()"
-      case "application/json"                  => "CodecFormat.Json()"
-      case "application/octet-stream"          => "CodecFormat.OctetStream()"
-      case "application/xml"                   => "CodecFormat.Xml()"
-      case "application/x-www-form-urlencoded" => "CodecFormat.XWwwFormUrlencoded()"
-      case "application/zip"                   => "CodecFormat.Zip()"
-      case o                                   => s"`${o}CodecFormat`()"
-    }
-    if (isEager) MappedContentType(eagerBody, if (contentType.startsWith("text/")) "String" else "Array[Byte]")
-    else {
-      val capability = capabilityImpl(streamingImplementation)
-      val tpe = capabilityType(streamingImplementation)
-      schema match {
-        case _: OpenapiSchemaString =>
-          MappedContentType(s"streamTextBody($capability)($streamingBody)", tpe)
-        case schema =>
-          val (outT, maybeInlineDefn) = schema match {
-            case st: OpenapiSchemaSimpleType =>
-              val (t, _) = mapSchemaSimpleTypeToType(st)
-              t -> None
-            case OpenapiSchemaArray(st: OpenapiSchemaSimpleType, _, _, _) =>
-              val (t, _) = mapSchemaSimpleTypeToType(st)
-              s"List[$t]" -> None
-            case OpenapiSchemaMap(st: OpenapiSchemaSimpleType, _, _) =>
-              val (t, _) = mapSchemaSimpleTypeToType(st)
-              s"Map[String, $t]" -> None
-            case o: OpenapiSchemaObject => inlineDefn(endpointName, position, o)
-            case x                      => bail(s"Can't create this param as output (found $x)")
-          }
-          MappedContentType(
-            s"streamBody($capability)(Schema.binary[$outT], $streamingBody)",
-            tpe,
-            inlineDefns = maybeInlineDefn,
-            inlineTypes = maybeInlineDefn.map(_ => outT).toSeq
-          )
-      }
-    }
-  }
 
 }
