+++ openapi-codegen/core/src/main/scala/sttp/tapir/codegen/endpoints/StreamingImplementation.scala
@@ -5,3 +5,9 @@ object Akka extends StreamingImplementation
 case class FS2(effectType: String = "cats.effect.IO") extends StreamingImplementation
 object Pekko extends StreamingImplementation
 object Zio extends StreamingImplementation
+
+
+object Position extends Enumeration {
+  val Request, Response, Err = Value
+  type Position = Value
+}
